Handshake, and NULL BYTE ending implemented

+++ client.c
@@ -20,29 +20,41 @@ void	listen(int signum, siginfo_t *server_info, void *context)
 	static int	handshake = 0;
 	static int	count = 1;
 	static int	str_index = 0;
-	char		c;
+	static int	bit_sent = 0;
 
 	(void) context;
 	if (signum == SIGUSR1 && handshake == 0)
 	{
 		handshake = 1;
-		ft_printf("handshake cplt");
 	}
-	if (signum == SIGUSR1 && str[str_index] != '\0' && count < 9 && handshake == 1)
+	if (signum == SIGUSR1 && str &&str[str_index] != '\0' && count < 9 && handshake == 1)
 	{
-		c = str[str_index];
-		if (c & (1 << (8-count)))
+		if (str[str_index] & (1 << (8-count)))
 			kill (server_info->si_pid,SIGUSR1);
 		else
 			kill(server_info->si_pid,SIGUSR2);
 		count ++;
+		bit_sent ++;
+	}
+	if (str && str[str_index] == '\0')
+	{
+		count = 0;
+		while (count < 8)
+		{
+			usleep(100);
+			kill(server_info->si_pid, SIGUSR2);
+			count ++;
+		}
+		ft_printf("%d bits sent\n",bit_sent);
+		str = NULL;
+
 	}
 	if (count == 9)
 	{
 		count = 1;
 		str_index ++;
 	}
-	usleep(50);
+
 }
 
 int	main(int argc, char **argv)
@@ -63,9 +75,9 @@ int	main(int argc, char **argv)
 	sigaction(SIGUSR2,&message_handler,NULL);
 	pid = (pid_t) ft_atoi(argv[1]);
 	talk(pid);
-	while (1)
+	while (str != NULL)
 	{
-		usleep(500);
+		usleep(50);
 	}
 	return (0);
 }
